fix: оптимизирована ручка initgame

+++ models/upgrade.go
@@ -7,7 +7,7 @@ type Upgrade struct {
 	Name           string		   `json:"name"`
 	IconName       string		   `json:"icon_name"`
 	UpgradeType    UpgradeType     `json:"upgrade_type" gorm:"type:upgrade_type"`
-	PriceFactor    float32		   `json:"price_factor"`
+	PriceFactor    float64		   `json:"price_factor"`
 	Price          uint			   `json:"price"`
 	AccessLevel    uint			   `json:"access_level"`
 	Boost          Boost           `json:"boost" gorm:"foreignKey:UpgradeID"`
